backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
werkzeug==3.0.1
watchdog==3.0.0 